cairo: moved definitions of cairopp functions to source files
